requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0
Werkzeug==3.0.1

# Astrology Calculations
pyswisseph==2.10.3.2

# AI/ML & APIs
openai>=1.50.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dotenv==1.0.0
requests==2.31.0

# Redis & Session
redis>=5.0.0
Flask-Session>=0.5.0


# Firebase Admin SDK (Push Notifications)
firebase-admin>=6.4.0

# Deployment (Vercel uses its own runtime)
serverless-wsgi==3.0.2
